requirements.txt
# pre-commit
fastapi
llama-index
llama-index-llms-openai
llama-index-llms-huggingface
llama-index-embeddings-huggingface
loguru
qdrant-client
llama-index-vector-stores-qdrant
uvicorn
# spacy
# scrapy
# playwright
# scrapy-playwright
python-dotenv
rootutils
langchain
langchain-openai
torch
transformers
